+++ controllers/NsetiController.php
@@ -150,7 +150,9 @@ class NsetiController extends Controller
     {
         $file = new Nseti();
         $model = $file->findOne($id);
-        unlink('uploads/files/'.$model->params);
+        if ($model->params) {
+            unlink('uploads/files/' . $model->params);
+        }
         $this->findModel($id)->delete();
 
 //unlink('/path/to/file');
